make by45_p21_w match the extern use in byvidpin.c, related to #451

+++ src/wpc/by35snd.c
@@ -736,7 +736,7 @@ static WRITE_HANDLER(cs_ctrl_w) {
   cpu_set_irq_line(cslocals.brdData.cpuNo, M6803_TIN_LINE, (data & 1) ? ASSERT_LINE : CLEAR_LINE);
 }
 
-void by45_p21_w(int data)
+void by45_p21_w(UINT8 data)
 {
 	cslocals.p21 = data ? 1 : 0;
 }
